ai-service/requirements.txt
# AI服务依赖 (使用版本范围，兼容LoomaCRM)
sanic>=23.12.0,<24.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
aiohttp>=3.9.0,<4.0.0
asyncpg>=0.29.0,<1.0.0
aiomysql>=0.2.0,<1.0.0
httpx>=0.25.0,<1.0.0

# AI/ML相关 (轻量级核心版本，避免内存不足)
# transformers>=4.30.2,<5.0.0  # 暂时禁用，内存不足
# torch>=2.0.0,<3.0.0  # 暂时禁用，内存不足
numpy>=1.24.3,<2.0.0
# pandas>=2.1.0,<3.0.0  # 暂时禁用，内存不足
# sentence-transformers>=2.2.0,<3.0.0  # 暂时禁用，内存不足
# scikit-learn>=1.3.0,<2.0.0  # 暂时禁用，内存不足（导致OOM）
# huggingface-hub>=0.16.0,<1.0.0  # 暂时禁用，内存不足
# 修复PyYAML版本冲突
PyYAML>=6.0.0,<7.0.0

# 向量化和嵌入相关
# fastembed==0.7.3  # 暂时移除，与huggingface-hub版本冲突
pgvector>=0.4.0,<1.0.0

# 数据库连接 (使用版本范围)
mysql-connector-python>=8.2.0,<10.0.0
pymysql>=1.1.0,<2.0.0

# 认证和权限 (使用版本范围)
PyJWT>=2.8.0,<3.0.0
cryptography>=41.0.0,<47.0.0

# 日志和监控 (使用版本范围)
structlog>=23.2.0,<24.0.0

# 开发工具 (轻量级版本，避免内存不足)
# pytest>=7.4.0,<9.0.0  # 暂时禁用，内存不足
# pytest-asyncio>=0.21.0,<2.0.0  # 暂时禁用，内存不足
# black>=23.12.0,<26.0.0  # 暂时禁用，内存不足
# flake8>=6.1.0,<8.0.0  # 暂时禁用，内存不足
